# fursa-ai/examples/openapi.json
{"openapi":"3.1.0","info":{"title":"Fursa AI Pipeline","description":"RAG + reasoning services for Fursa, an AI-Powered Workforce Readiness Ecosystem for Saudi Arabia (ITU-T Y.3172 aligned).","version":"0.1.0"},"paths":{"/api/health":{"get":{"tags":["system"],"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Api Health Get"}}}}}}},"/api/kb/stats":{"get":{"tags":["system"],"summary":"Kb Stats","operationId":"kb_stats_api_kb_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Kb Stats Api Kb Stats Get"}}}}}}},"/api/kb/search":{"get":{"tags":["system"],"summary":"Kb Search","operationId":"kb_search_api_kb_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","description":"Semantic search with no LLM — useful for debugging.","title":"Q"},"description":"Semantic search with no LLM — useful for debugging."},{"name":"top_k","in":"query","required":false,"schema":{"type":"integer","maximum":25,"minimum":1,"default":5,"title":"Top K"}},{"name":"raw","in":"query","required":false,"schema":{"type":"boolean","description":"True = pure top-k by distance. False (default) = the production behaviour, capped at 2 chunks per document. Note /api/ask-kb additionally splits multi-part questions into per-clause queries.","default":false,"title":"Raw"},"description":"True = pure top-k by distance. False (default) = the production behaviour, capped at 2 chunks per document. Note /api/ask-kb additionally splits multi-part questions into per-clause queries."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Kb Search Api Kb Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/match":{"post":{"tags":["ml"],"summary":"Api Match","operationId":"api_match_api_match_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/recommend-courses":{"post":{"tags":["ml"],"summary":"Api Recommend Courses","operationId":"api_recommend_courses_api_recommend_courses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendCoursesRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendCoursesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/readiness-score":{"post":{"tags":["ml"],"summary":"Api Readiness Score","operationId":"api_readiness_score_api_readiness_score_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadinessScoreRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadinessScoreResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/adaptive-path":{"post":{"tags":["ml"],"summary":"Api Adaptive Path","operationId":"api_adaptive_path_api_adaptive_path_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdaptiveLearningRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdaptiveLearningResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/skill-trends":{"get":{"tags":["ml"],"summary":"Api Skill Trends Get","operationId":"api_skill_trends_get_api_skill_trends_get","parameters":[{"name":"horizon_months","in":"query","required":false,"schema":{"type":"integer","maximum":60,"minimum":1,"default":12,"title":"Horizon Months"}},{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"region","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Region"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillTrendsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["ml"],"summary":"Api Skill Trends Post","operationId":"api_skill_trends_post_api_skill_trends_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillTrendsRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillTrendsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/ask-kb":{"post":{"tags":["rag"],"summary":"Api Ask Kb","operationId":"api_ask_kb_api_ask_kb_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AskKBRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AskKBResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AdaptiveLearningRequest":{"properties":{"student":{"$ref":"#/components/schemas/StudentProfile"},"progress_history":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Progress History","description":"Chronological learning events. Any shape; useful keys: course_id, title, skill, status, score, attempts, started_at, completed_at, time_spent_hours, dropped."},"current_path":{"items":{"type":"string"},"type":"array","title":"Current Path"},"target_role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role"}},"type":"object","required":["student","progress_history"],"title":"AdaptiveLearningRequest"},"AdaptiveLearningResponse":{"properties":{"status":{"type":"string","enum":["on_track","at_risk","struggling","ready_to_accelerate"],"title":"Status"},"diagnosis":{"type":"string","title":"Diagnosis"},"alternative_paths":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Alternative Paths"},"recommended_adjustments":{"items":{"type":"string"},"type":"array","title":"Recommended Adjustments"},"motivation_note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Motivation Note"}},"additionalProperties":true,"type":"object","required":["status","diagnosis"],"title":"AdaptiveLearningResponse"},"AskKBRequest":{"properties":{"question":{"type":"string","title":"Question"},"top_k":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Top K"},"kb_ids":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Kb Ids","description":"Restrict retrieval to specific documents, e.g. ['KB-010']."}},"type":"object","required":["question"],"title":"AskKBRequest"},"AskKBResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"citations":{"items":{"type":"string"},"type":"array","title":"Citations"},"confidence":{"type":"string","enum":["high","medium","low"],"title":"Confidence","default":"medium"},"grounded":{"type":"boolean","title":"Grounded","default":true},"follow_up_questions":{"items":{"type":"string"},"type":"array","title":"Follow Up Questions"},"sources":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Sources"}},"additionalProperties":true,"type":"object","required":["answer"],"title":"AskKBResponse"},"Certification":{"properties":{"name":{"type":"string","title":"Name"},"issuer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Issuer"},"year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"},"verified":{"type":"boolean","title":"Verified","default":false}},"type":"object","required":["name"],"title":"Certification"},"Course":{"properties":{"course_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Course Id"},"title":{"type":"string","title":"Title"},"provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Provider","description":"e.g. Doroob, TVTC, Coursera"},"skills_taught":{"items":{"type":"string"},"type":"array","title":"Skills Taught"},"duration_hours":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Duration Hours"},"level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Level"},"cost_sar":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cost Sar"},"is_free":{"type":"boolean","title":"Is Free","default":false},"certification_awarded":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Certification Awarded"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"}},"type":"object","required":["title"],"title":"Course"},"CourseRecommendation":{"properties":{"course_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Course Id"},"title":{"type":"string","title":"Title"},"rank":{"type":"integer","title":"Rank"},"closes_gaps":{"items":{"type":"string"},"type":"array","title":"Closes Gaps"},"impact_score":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Impact Score"},"rationale":{"type":"string","title":"Rationale"},"estimated_readiness_gain":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Estimated Readiness Gain"},"sequence_note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sequence Note"}},"additionalProperties":true,"type":"object","required":["title","rank","impact_score","rationale"],"title":"CourseRecommendation"},"Experience":{"properties":{"title":{"type":"string","title":"Title"},"organization":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Organization"},"months":{"type":"integer","title":"Months","default":0},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["title"],"title":"Experience"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobRequirements":{"properties":{"job_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Id"},"title":{"type":"string","title":"Title"},"employer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Employer"},"sector":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"required_skills":{"items":{"type":"string"},"type":"array","title":"Required Skills"},"preferred_skills":{"items":{"type":"string"},"type":"array","title":"Preferred Skills"},"required_certifications":{"items":{"type":"string"},"type":"array","title":"Required Certifications"},"min_experience_months":{"type":"integer","title":"Min Experience Months","default":0},"education_requirement":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Education Requirement"},"is_entry_level":{"type":"boolean","title":"Is Entry Level","default":true}},"type":"object","required":["title"],"title":"JobRequirements"},"MatchRequest":{"properties":{"job":{"$ref":"#/components/schemas/JobRequirements"},"student":{"$ref":"#/components/schemas/StudentProfile"},"explain":{"type":"boolean","title":"Explain","description":"Include the natural-language rationale.","default":true}},"type":"object","required":["job","student"],"title":"MatchRequest"},"MatchResponse":{"properties":{"match_score":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Match Score"},"verdict":{"type":"string","enum":["strong_match","promising","developing","not_yet_ready"],"title":"Verdict"},"explanation":{"type":"string","title":"Explanation"},"matched_skills":{"items":{"type":"string"},"type":"array","title":"Matched Skills"},"gaps":{"items":{"$ref":"#/components/schemas/SkillGap"},"type":"array","title":"Gaps"},"breakdown":{"additionalProperties":true,"type":"object","title":"Breakdown"},"recommended_next_actions":{"items":{"type":"string"},"type":"array","title":"Recommended Next Actions"},"fairness_note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fairness Note"}},"additionalProperties":true,"type":"object","required":["match_score","verdict","explanation"],"title":"MatchResponse"},"Project":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"skills_used":{"items":{"type":"string"},"type":"array","title":"Skills Used"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"}},"type":"object","required":["name"],"title":"Project"},"ReadinessScoreRequest":{"properties":{"student":{"$ref":"#/components/schemas/StudentProfile"},"target_role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role"},"job_market_context":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Market Context","description":"Optional demand signals to weight the score against."}},"type":"object","required":["student"],"title":"ReadinessScoreRequest"},"ReadinessScoreResponse":{"properties":{"readiness_score":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Readiness Score"},"band":{"type":"string","enum":["early_stage","developing","job_ready","highly_competitive"],"title":"Band"},"breakdown":{"additionalProperties":true,"type":"object","title":"Breakdown"},"strengths":{"items":{"type":"string"},"type":"array","title":"Strengths"},"weaknesses":{"items":{"type":"string"},"type":"array","title":"Weaknesses"},"highest_impact_next_step":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Highest Impact Next Step"},"explanation":{"type":"string","title":"Explanation"}},"additionalProperties":true,"type":"object","required":["readiness_score","band","explanation"],"title":"ReadinessScoreResponse"},"RecommendCoursesRequest":{"properties":{"skill_gaps":{"items":{"type":"string"},"type":"array","title":"Skill Gaps","description":"Gap skill names, e.g. from /api/match"},"course_catalog":{"items":{"$ref":"#/components/schemas/Course"},"type":"array","title":"Course Catalog"},"student":{"anyOf":[{"$ref":"#/components/schemas/StudentProfile"},{"type":"null"}]},"target_role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role"},"max_recommendations":{"type":"integer","title":"Max Recommendations","default":5},"constraints":{"additionalProperties":true,"type":"object","title":"Constraints","description":"Free-form, e.g. {'max_hours_per_week': 6, 'budget_sar': 0}"}},"type":"object","required":["skill_gaps"],"title":"RecommendCoursesRequest"},"RecommendCoursesResponse":{"properties":{"recommendations":{"items":{"$ref":"#/components/schemas/CourseRecommendation"},"type":"array","title":"Recommendations"},"uncovered_gaps":{"items":{"type":"string"},"type":"array","title":"Uncovered Gaps"},"learning_path_summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Learning Path Summary"}},"additionalProperties":true,"type":"object","title":"RecommendCoursesResponse"},"Skill":{"properties":{"name":{"type":"string","title":"Name"},"level":{"anyOf":[{"type":"string","enum":["beginner","intermediate","advanced","expert"]},{"type":"null"}],"title":"Level"},"years":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Years"},"verified":{"type":"boolean","title":"Verified","default":false}},"type":"object","required":["name"],"title":"Skill"},"SkillGap":{"properties":{"skill":{"type":"string","title":"Skill"},"importance":{"type":"string","enum":["critical","important","nice_to_have"],"title":"Importance","default":"important"},"student_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Student Level"},"required_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Required Level"},"closable_in_weeks":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Closable In Weeks"}},"additionalProperties":true,"type":"object","required":["skill"],"title":"SkillGap"},"SkillTrendsRequest":{"properties":{"job_postings":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Job Postings","description":"Historical postings. Useful keys: title, sector, posted_at (ISO date), region, required_skills[], count."},"horizon_months":{"type":"integer","title":"Horizon Months","default":12},"sector":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"},"region":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Region"}},"type":"object","required":["job_postings"],"title":"SkillTrendsRequest"},"SkillTrendsResponse":{"properties":{"trending_up":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Trending Up"},"trending_down":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Trending Down"},"stable":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Stable"},"emerging":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Emerging"},"horizon_months":{"type":"integer","title":"Horizon Months","default":12},"curriculum_implications":{"items":{"type":"string"},"type":"array","title":"Curriculum Implications"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"},"data_window":{"additionalProperties":true,"type":"object","title":"Data Window"}},"additionalProperties":true,"type":"object","title":"SkillTrendsResponse"},"StudentProfile":{"properties":{"student_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Student Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"target_role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Role","description":"e.g. 'Data Analyst'"},"education_level":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Education Level"},"major":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Major"},"university":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"University"},"graduation_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Graduation Year"},"skills":{"items":{"$ref":"#/components/schemas/Skill"},"type":"array","title":"Skills"},"soft_skills":{"items":{"type":"string"},"type":"array","title":"Soft Skills"},"certifications":{"items":{"$ref":"#/components/schemas/Certification"},"type":"array","title":"Certifications"},"experience":{"items":{"$ref":"#/components/schemas/Experience"},"type":"array","title":"Experience"},"projects":{"items":{"$ref":"#/components/schemas/Project"},"type":"array","title":"Projects"},"languages":{"items":{"type":"string"},"type":"array","title":"Languages"},"preferred_locations":{"items":{"type":"string"},"type":"array","title":"Preferred Locations"}},"type":"object","title":"StudentProfile","description":"The Skills Passport."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}